Resolve issues with UserSetsControllerMetrics concern
