+++ app/controllers/supplejack_api/user_sets_controller.rb
@@ -7,14 +7,12 @@
 
 module SupplejackApi
   class UserSetsController < ApplicationController
+    include Concerns::UserSetsControllerMetrics
 
     respond_to :json
 
-    before_filter :find_user_set, only: [:update, :destroy]
+    prepend_before_filter :find_user_set, only: [:update, :destroy]
     before_filter :authenticate_admin!, only: [:admin_index, :public_index]
-    before_action :set_existing_user_set_items, only: :update
-    after_action :create_set_interaction, only: :update
-    after_action :create_set_record_view, only: :show
 
     def index
       @user_sets = current_user.user_sets
@@ -78,31 +76,6 @@ module SupplejackApi
     end
 
     private
-
-    def set_existing_user_set_items
-      return unless @user_set
-      
-      @existing_user_set_items = @user_set.set_items.dup
-    end
-
-    def create_set_interaction
-      return unless params[:set].key? :records
-
-      record_ids = params[:set][:records].map{|x| x[:record_id]}
-      new_record_ids = record_ids.reject{|id| @existing_user_set_items.any?{|set_item| set_item.record_id.to_s == id}}
-      display_collections = SupplejackApi::Record.find_multiple(new_record_ids).map(&:display_collection)
-
-      display_collections.each{|dc| InteractionModels::Set.create(interaction_type: :creation, facet: dc)}
-    end
-
-    def create_set_record_view
-      return unless @user_set
-
-      SupplejackApi::InteractionModels::Record.create_user_set(
-        @user_set, 
-        params[:request_logger_field]
-      ) if params[:request_logger]
-    end
 
     def serializable_array(user_sets, options={})
       options.reverse_merge!(root: false, items: false, user: false, total: false)
