+++ config/initializers/interaction_updaters.rb
@@ -1,0 +1,3 @@
+[SupplejackApi::InteractionUpdaters::SetMetrics, SupplejackApi::InteractionUpdaters::UsageMetrics].each do |x|
+  SupplejackApi::InteractionMetricsWorker.register_interaction_updater(x)
+end
